+++ philo/philo.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   philo.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: imunaev- <imunaev-@studen.hive.fi>         +#+  +:+       +#+        */
+/*   By: imunaev- <imunaev-@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/14 00:04:08 by imunaev-          #+#    #+#             */
-/*   Updated: 2025/02/20 14:52:53 by imunaev-         ###   ########.fr       */
+/*   Updated: 2025/02/20 21:38:22 by imunaev-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,33 +23,33 @@
 
 typedef struct s_sim
 {
-	long			ph_count;
+	int				ph_count;
 	bool			all_dead;
 	bool			all_satiated;
-	long			num_meals_to_eat;
-	long			*meal_count;
+	int				num_meals_to_eat;
+	int				*meal_count;
 	pthread_t		*ph_threads;
 	pthread_t		monitor_thread;
 	pthread_mutex_t	*mtx_forks;
 	pthread_mutex_t	mtx_is_dead;
 	pthread_mutex_t	mtx_log;
 	pthread_mutex_t	mtx_last_meal_time;
-	pthread_mutex_t	mtx_meal_count;    
+	pthread_mutex_t	mtx_meal_count;
 }	t_sim;
 
 typedef struct s_ph
 {
 	size_t		index;
 	uint64_t	last_meal_time;
-	long		time_to_die;
-	long		time_to_eat;
-	long		time_to_sleep;
+	int			time_to_die;
+	int			time_to_eat;
+	int			time_to_sleep;
 	bool		is_dead;
 	t_sim		*sim;
 }	t_ph;
 
 bool		is_input_valid(int ac, char **av);
-long		ft_atol(const char *str);
+int			ft_atol(const char *str);
 void		print_err(char *msg);
 void		*philo(void *arg);
 void		think(t_ph *ph);
